refactor: 이벤트 공유 컴포넌트 분리 및 FinalResult UX 개선
