+++ client/src/pages/RushGame/index.tsx
@@ -1,10 +1,7 @@
 import { useEffect } from "react";
-import { motion } from "framer-motion";
 import { useCookies } from "react-cookie";
 import { useLoaderData } from "react-router-dom";
-import CTAButton from "@/components/CTAButton";
 import { CARD_PHASE } from "@/constants/Rush/rushCard.ts";
-import { ASCEND, SCROLL_MOTION } from "@/constants/animation.ts";
 import { COOKIE_KEY } from "@/constants/cookie.ts";
 import CardOptions from "@/features/RushGame/RushGameSections/CardOptions.tsx";
 import Countdown from "@/features/RushGame/RushGameSections/Countdown.tsx";
@@ -15,9 +12,7 @@ import { useFetchRushUserParticipationStatus } from "@/hooks/RushGame/useFetchRu
 import { useFetchTodayRushEvent } from "@/hooks/RushGame/useFetchTodayRushEvent.ts";
 import useSetGamePhase from "@/hooks/RushGame/useSetGamePhase.ts";
 import { useBlockNavigation } from "@/hooks/useBlockNavigation.ts";
-import useToast from "@/hooks/useToast.tsx";
 import { GetTotalRushEventsResponse } from "@/types/rushApi.ts";
-import { writeClipboard } from "@/utils/writeClipboard.ts";
 
 export default function RushGame() {
     const [cookies] = useCookies([COOKIE_KEY.ACCESS_TOKEN]);
@@ -26,16 +21,11 @@ export default function RushGame() {
     );
     const { getTodayRushEvent } = useFetchTodayRushEvent();
     const gameState = useRushGameStateContext();
-    const { showToast, ToastComponent } = useToast("🔗 링크가 복사되었어요!");
     const { getRushUserParticipationStatus, userParticipatedStatus } =
         useFetchRushUserParticipationStatus();
 
     const rushData = useLoaderData() as GetTotalRushEventsResponse;
     useSetGamePhase(rushData);
-
-    const handleClickShareButton = () => {
-        writeClipboard(import.meta.env.VITE_RUSH_URL, showToast);
-    };
 
     useEffect(() => {
         getTodayRushEvent(cookies[COOKIE_KEY.ACCESS_TOKEN]);
@@ -43,6 +33,7 @@ export default function RushGame() {
     }, []);
 
     const renderRushGameContent = () => {
+        if (gameState.phase === null) return null;
         switch (gameState.phase) {
             case CARD_PHASE.NOT_STARTED:
                 return <Countdown />;
@@ -65,17 +56,6 @@ export default function RushGame() {
     return (
         <section className="h-screen bg-n-white flex flex-col gap-8 justify-center items-center">
             {renderRushGameContent()}
-            <motion.div
-                className="flex flex-col justify-center items-center gap-4 my-3"
-                {...SCROLL_MOTION(ASCEND)}
-            >
-                <p className="h-body-2-regular text-n-neutral-500">
-                    우리 편에 투표할 친구를 불러오세요!
-                </p>
-                <CTAButton label="이벤트 링크 공유" onClick={handleClickShareButton} />
-            </motion.div>
-
-            {ToastComponent}
         </section>
     );
 }
